record_function now using singleton and in its own file decorators.py
