+++ agentops/__init__.py
@@ -4,3 +4,4 @@ from .client import Client
 from .event import ActionEvent, LLMEvent, ToolEvent, ErrorEvent
 from .logger import AgentOpsLogger
 from .enums import Models
+from .decorators import record_function
