+++ tests/openai_handlers/test_llm_tracker_lt_1_sync.py
@@ -1,10 +1,10 @@
 import openai
 from openai import ChatCompletion
 import agentops
+from agentops import record_function
 from packaging.version import parse
 from importlib import import_module
 import sys
-
 
 ao_client = agentops.Client()
 
@@ -17,7 +17,7 @@ if api in sys.modules:
             print('openai version: ', module_version)
 
 
-@ao_client.record_function('sample function being recorded')
+@record_function('sample function being recorded')
 def call_openai():
 
     message = [{"role": "user", "content": "Hello"},
